fix(@ballerine/blocks): fixed types resolving to any
