+++ apps/backoffice-v2/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ballerine/backoffice-v2",
-  "version": "0.5.46",
+  "version": "0.5.47",
   "description": "Ballerine - Backoffice",
   "homepage": "https://github.com/ballerine-io/ballerine",
   "repository": {
@@ -50,7 +50,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@ballerine/blocks": "0.1.25",
+    "@ballerine/blocks": "0.1.26",
     "@ballerine/common": "0.7.41",
     "@ballerine/ui": "^0.3.28",
     "@ballerine/workflow-browser-sdk": "0.5.41",
