+++ apps/kyb-app/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@ballerine/kyb-app",
   "private": true,
-  "version": "0.1.41",
+  "version": "0.1.42",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -14,7 +14,7 @@
     "test:dev": "vitest"
   },
   "dependencies": {
-    "@ballerine/blocks": "0.1.25",
+    "@ballerine/blocks": "0.1.26",
     "@ballerine/common": "^0.7.41",
     "@ballerine/ui": "0.3.28",
     "@ballerine/workflow-browser-sdk": "0.5.41",
