+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/blocks
 
+## 0.1.26
+
+### Patch Changes
+
+- Fixed types resolving to any
+
 ## 0.1.25
 
 ### Patch Changes
