+++ packages/blocks/package.json
@@ -2,7 +2,7 @@
   "private": false,
   "name": "@ballerine/blocks",
   "author": "Ballerine <dev@ballerine.com>",
-  "version": "0.1.25",
+  "version": "0.1.26",
   "description": "blocks",
   "module": "./dist/esm/index.js",
   "main": "./dist/cjs/index.js",
