+++ packages/blocks/src/BlocksComponent.stories.tsx
@@ -1,5 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { BlocksComponent, BlocksProps, CellsMap, createBlocks, ExtractCellProps } from './';
+import { createBlocks } from '@/blocks';
+import { BlocksComponent } from '@/BlocksComponent';
+import type { BlocksProps, CellsMap, ExtractCellProps } from '@/blocks';
 import { ComponentProps, FunctionComponent } from 'react';
 
 const meta = {
