+++ packages/blocks/src/BlocksComponent.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { ForwardedRef, forwardRef, Fragment } from 'react';
-import type { BlocksProps, Cells } from './types';
+
+import { BlocksProps, Cells } from '@/blocks';
 
 export const BlocksComponent = forwardRef(
   <TCell extends Cells>(
