+++ packages/blocks/src/index.ts
@@ -1,3 +1,2 @@
-export * from './types';
 export * from './blocks';
 export * from './BlocksComponent';
